[ADD] Usable as package and basic cli
