+++ RELEASE.md
@@ -1,0 +1,10 @@
+New features in the latest update
+=====================================
+
+
+0.1.0 [Feb 18th, 2019]:
+---
+
+* First release of the tool!
+* Usable as a cli command!
+* Function as describe in README!
