+++ giteasychangelog/__init__.py
@@ -5,3 +5,4 @@
 __author__ = """Eric Horvat"""
 __email__ = 'eric.nahuel.horvat@gmail.com'
 __version__ = '0.1.0'
+__main__ = 'cli'
