+++ requirements.txt
@@ -1,1 +1,2 @@
 packaging >= 18.0
+click >= 6.0
